+++ api/types/courseData.ts
@@ -1,0 +1,6 @@
+export default interface CourseData {
+    code: string;
+    name: string;
+    description: string;
+    educator: string;
+  }
