+++ client/package.json
@@ -3,7 +3,6 @@
   "version": "0.1.0",
   "private": true,
   "proxy": "http://localhost:8080/",
-  "secure": false,
   "dependencies": {
     "@material-ui/core": "^4.11.0",
     "@material-ui/icons": "^4.9.1",
@@ -15,6 +14,7 @@
     "@types/react": "^16.9.53",
     "@types/react-dom": "^16.9.8",
     "@types/react-router-dom": "^5.1.6",
+    "axios": "^0.21.0",
     "dotenv": "^8.2.0",
     "firebase": "^8.0.2",
     "react": "^17.0.1",
@@ -52,7 +52,6 @@
     ]
   },
   "devDependencies": {
-    "axios": "^0.21.0",
     "prettier": "^2.1.2"
   }
 }
